// ui/src/features/rum/api/rum-sessions.generated.ts
export type RumSessionsQuery = {
  readonly rumSessions: ReadonlyArray<{
    readonly sessionId: string
    readonly service: string
    readonly startNanos: string
    readonly endNanos: string
    readonly spanCount: number
    readonly traceCount: number
    readonly viewCount: number
    readonly vitalCount: number
    readonly errorCount: number
    readonly hasError: boolean
  }>
}